Resolve an issue with irc_nick_sync in poco-irc-state
There was bug with the first irc_nick_sync event triggered when a user
joins a channel and that user is unknown to the poco.

Due to unfortunate variable naming the NICK and CHANNEL name were being
reported in the wrong order in the irc_nick_sync event.

I have changed variable names to clarify the situation.