Store the real nick for channel members after a NAMES reply
We were adding nicks to channel lists after NAMES replies but not
storing the "real" nick in the state. This meant that if someone
requested a list of channel members before the channel was synced,
some of the entires would be undefined. Any queries made about the
nicks would then fail:

    for my $member ($irc->channel_list("#chan") {
        # $member might be undef
        print "yay" if $irc->is_channel_operator($member);
    }