Don't queue messages if $self->{flood}
I think this should fix #10.

If `$self->{flood}` is set, then just send the message - even if there
is currently a queue (because e.g. we wanted to send stuff while we were
disconnected, so queued it).

Otherwise, for a reasonably chatty bot, once it has disconnected once,
and one or more messages got queued in the meantime, all future messages
will go into the queue too, ignoring `$self->{flood}` - and that will
remain until such time as the queue manages to empty, which for a chatty
bot might not happen.